# src/main/java/com/banking/dto/auth/LoginResponse.java
package com.banking.dto.auth;

import com.banking.dto.user.UserDTO;
import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class LoginResponse {
    private String token;
    private UserDTO user;
}