# src/main/java/com/banking/dto/user/UserDTO.java
package com.banking.dto.user;

import com.banking.entity.User;
import com.banking.enums.UserRole;
import lombok.Data;

import java.time.LocalDate;

@Data
public class UserDTO {
    private Long id;
    private String username;
    private String firstName;
    private String lastName;
    private String email;
    private String phoneNumber;
    private String bsn;
    private LocalDate dateOfBirth;
    private UserRole role;
    private boolean approved;

    public static UserDTO from(User user) {
        UserDTO dto = new UserDTO();
        dto.setId(user.getId());
        dto.setUsername(user.getUsername());
        dto.setFirstName(user.getFirstName());
        dto.setLastName(user.getLastName());
        dto.setEmail(user.getEmail());
        dto.setPhoneNumber(user.getPhoneNumber());
        dto.setBsn(user.getBsn());
        dto.setDateOfBirth(user.getDateOfBirth());
        dto.setRole(user.getRole());
        dto.setApproved(user.isApproved());
        return dto;
    }
}